Move announcement handling into its own module
All announcement logic lived spread across the player and streams
controllers, with the two of them sharing a plain dict that either could
mutate.

- Move announcement playback out of the player controller into an
  AnnouncementsMixin, next to the existing protocol linking mixin.
- Let the announcement renderer own which announcement each player is
  playing, so registering and releasing one is a single paired call
  instead of two registries torn down from two different controllers.
- get_announcement_url() now only builds the url instead of registering
  an announcement as a side effect.
